custom Docker image work for CI
align the CI workflow with the Docker based development workflow and
roll in some tweaks to both workflows along the way

* created a .docker dir with some Docker specific content
   * moved the ruby_installer.rb script here
   * reworked the Ruby/Rails mapper to be able to live here within a
     container
* add a helper method to the Docker helper class to set environment
  variables when shelling out to Bundler to set things needed for
  older Rubies (2.3 and 2.4)
* use v0.3.21, the latest v0.3.x version of mysql2 for the v0.3.x tests
* annotate Dockerfile
* use the latest JRuby, v9.3.3.0
* defer CI RubyGem bundling to ruby/setup-ruby [sic]
* in CI, run MySQL server as a service separate from the app
* Git ignore exuberant C tags

+++ lib/tasks/multiverse.rb
@@ -44,7 +44,10 @@ namespace :test do
   namespace :multiverse do
     def remove_local_multiverse_databases
       list_databases_command = %(echo "show databases" | mysql -u root)
-      databases = `#{list_databases_command}`.chomp!.split("\n").select { |s| s =~ /multiverse/ }
+      all_databases = `#{list_databases_command}`
+      raise "Failed to obtain a list of databases by running '#{list_databases_command}'" if all_databases == ""
+
+      databases = all_databases.chomp!.split("\n").select { |s| s =~ /multiverse/ }
       databases.each do |database|
         puts "Dropping #{database}"
         `echo "drop database #{database}" | mysql -u root`
